move summary statistic functions
